Implement database abstraction layer
