+++ backend/test/test_oracle.py
@@ -1,29 +1,33 @@
 import unittest
-from backend import oracle, dynamodb
+from backend import oracle
+from backend.db import db
+
+db = db.get_db()
 
 
 class TestOracle(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        dynamodb.create_table()
+        print(db.get_config())
+        db.create_table()
 
     def test_orders(self):
         orders1 = oracle.handler({"status": -1}, None)
         self.assertEqual('{"orders": 0}', orders1)
-        self.assertEqual(1, dynamodb.number_of_items_in_table())
+        self.assertEqual(1, db.number_of_items_in_table())
 
         orders2 = oracle.handler({"status": 0}, None)
         self.assertEqual('{"orders": 1}', orders2)
-        self.assertEqual(2, dynamodb.number_of_items_in_table())
+        self.assertEqual(2, db.number_of_items_in_table())
 
         orders3 = oracle.handler({"status": 1}, None)
         self.assertEqual('{"orders": -1}', orders3)
-        self.assertEqual(3, dynamodb.number_of_items_in_table())
+        self.assertEqual(3, db.number_of_items_in_table())
 
     @classmethod
     def tearDownClass(cls):
-        dynamodb.delete_table()
+        db.delete_table()
 
 
 if __name__ == '__main__':
